+++ RELEASE_NOTES.md
@@ -10,6 +10,8 @@
 * [#692] Added Spell Importer macro
 * [#695] Add ability to have Effects on Boon items
 * [#697] Updated the formating of monster descriptions created with the monster importer.
+* [#704] Open character sheet after it has been created by the character creator
+* [#705] Character creator should be usable by players without giving them full permission to create actors
 
 ## Feature Notes
 A new Character Generator is now available via the actors tab. The generator can be used to create new characters randomly or via selection. Features include: